pyproject.toml
[tool.poetry]
name = "project-shopply"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
